backend/requirements.txt
Django>=5.1.2,<5.2
djangorestframework>=3.15.2,<3.16
django-extensions>=3.2.3,<3.3
psycopg2>=2.9.9,<3.0
drf-spectacular>=0.27.2,<0.28
djangorestframework-gis>=1.1,<1.2